requirements.txt
allure_python_commons~=2.12.0
allure-pytest~=2.12.0
assertpy~=1.1
Cerberus~=1.3.4
names~=0.3.0
pytest~=7.2.0
requests~=2.28.1
